+++ src/g_game.c
@@ -62,7 +62,7 @@ gamestate_t gamestate = GS_NULL;
 UINT8 ultimatemode = false;
 
 boolean botingame;
-UINT8 botskin;
+UINT16 botskin;
 UINT8 botcolor;
 
 JoyType_t Joystick;
